+++ doc/release_notes.rst
@@ -1,6 +1,9 @@
 Release notes
 =============
 
+* load_irf moved from bin/xpobssim.py to irf/__init__.py, so that it can be
+  reused.
+
 
 *ximpol (0.10.1) - Tue, 19 Jan 2016 16:41:33 -0800*